new Delete and Restore user

+++ src/assets/pages/adminPanel/AdminUserList.jsx
@@ -1,32 +1,61 @@
-import { API_PATH_USERS } from "../../common/constants/api_path.constants";
+import { API_PATH_RESTOREUSER, API_PATH_USERS } from "../../common/constants/api_path.constants";
 
 const AdminUserList = ({id, name, email, role, createdAt, deletedAt, onUserDelete }) => {
+
+    console.log(deletedAt);
 
     const handleClick = async (e) =>
     {
         e.preventDefault()
 
-        try {
-            const response = await fetch(`${API_PATH_USERS}/${id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-                    // 'Authorization': `Bearer ${localStorage.getItem('token')}`
-                },
-                body: JSON.stringify({ id }),
-            });
-
-            console.log(response);
-
-            if (!response.ok) {
-                throw new Error("Datos invalidos");
+        if(deletedAt)
+        {
+            try {
+                const response = await fetch(`${API_PATH_RESTOREUSER}/${id}`, {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json'
+                        // 'Authorization': `Bearer ${localStorage.getItem('token')}`
+                    },
+                    body: JSON.stringify({ id }),
+                });
+    
+                console.log(response);
+    
+                if (!response.ok) {
+                    throw new Error("Datos invalidos");
+                }
+    
+                onUserDelete();
+            } catch (error) {
+                console.error(error);
             }
-
-            onUserDelete();
-
-        } catch (error) {
-            console.error(error);
         }
+        else
+        {
+            try {
+                const response = await fetch(`${API_PATH_USERS}/${id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                        // 'Authorization': `Bearer ${localStorage.getItem('token')}`
+                    },
+                    body: JSON.stringify({ id }),
+                });
+    
+                console.log(response);
+    
+                if (!response.ok) {
+                    throw new Error("Datos invalidos");
+                }
+    
+                onUserDelete();
+    
+            } catch (error) {
+                console.error(error);
+            }
+        }
+        
     }
 
     return (
@@ -36,9 +65,9 @@ const AdminUserList = ({id, name, email, role, createdAt, deletedAt, onUserDelet
             <p>{role}</p>
             <p>{email}</p>
             <p>{createdAt}</p>
-            <p>{deletedAt}</p>
+            <p>{deletedAt ? deletedAt : '-'}</p>
 
-            <button onClick={handleClick}>DELETE</button>
+            <button onClick={handleClick}>{deletedAt ? 'RESTORE' : 'DELETE'}</button>
         </div>
     )
 }
